Split up awards tests
